requirements.txt
fastapi
uvicorn[standard]
mangum
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz
pandas
python-dotenv
pydantic
httpx
